test LMID

+++ dlfcn.go
@@ -58,8 +58,8 @@ func Dlmopen(lmid int32, path string, mode int) (uintptr, error) {
 // dynamically loaded object referred to by handle.  This is used with Dlmopen
 // to load additional libraries into the namespace of this loaded library.
 // Note that GCC limits the total number of namespaces to 16.
-func DlinfoLMID(handle uintptr) (int32, error) {
-	var lmid = int32(0)
+func DlinfoLMID(handle uintptr) (int64, error) {
+	var lmid = int64(0)
 	fmt.Printf("handle %v req %v lmid %v lmid_ptr %v\n", handle, RTLD_DI_LMID, lmid, uintptr(unsafe.Pointer(&lmid)))
 	u := fnDlinfoLmid(handle, RTLD_DI_LMID, uintptr(unsafe.Pointer(&lmid)))
 	fmt.Printf("return u %v lmid %v\n", u, lmid)
